refactor: move `load_world_manifest` to utils
